Convert create-test-archives script from CJS to TypeScript
